+++ src/main/java/com/example/starhub/repository/PostRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface PostRepository extends JpaRepository<PostEntity, Long> {
+public interface PostRepository extends JpaRepository<PostEntity, Integer> {
     public List<PostListResponseDto> findAllByOrderByCreatedAtDesc();
 }
